Update README.md
Food Delivery Cost and Profitability Analysis is all about understanding the financial side of a food delivery businesswhere money is spent, how revenue is made, and how to improve profits. The goal is to spot opportunities to cut costs, boost income, and adjust pricing or commission strategies to make the business more sustainable.